[*]帧同步调试第一阶段
下一阶段是通过实现自定义Math类，同步计算所有浮点数等，是的底层计算在不同客户端是一样的。

+++ Assets/GameMain/Scripts/NetWork/SendManager.cs
@@ -12,7 +12,7 @@ namespace TankBattle {
             Send(mStartMoveReq);
         }
 
-        public static void SendChangeDir(float tAngle) {
+        public static void SendChangeDir(int tAngle) {
             ChangeDirReq mChangeDirReq = new ChangeDirReq();
 
             mChangeDirReq.UserId = "" + GameEntry.NetData.mUserData.UserId;
